+++ SpeedyWeather/src/parameterizations/general.jl
@@ -1,10 +1,3 @@
-"""$TYPEDEF
-Abstract type for all column-based parmaetrizations. Custom parametrizations are expected to 
-subtype this and implement the [`variables`](@ref), [`initialize!`](@ref), and [`parameterization!`](@ref) for it. In
-order to use the parameterization in a model, add it to the `parameterizations` of the model 
-at definition."""
-abstract type AbstractParameterization <: AbstractModelComponent end
-
 """Function that defines the actual parameterization of an `AbstractParameterization`.
 
 Takes in the index of the column `ij`, the `Variables` object `vars`, the